// src/types/candidate_personal_data_types.ts
export interface CandidatePersonalDataTypes {
    uid?: string;
    id?: number;
    picture?: null;
    fullname?: string;
    number?: string;
    email?: string;
    address?: string;
    position?: string;
    level?: string;
    time_availabled?: string;
    estimat_salary?: string;
    experience_years?: string;
    current_company?: string;
    technlogies?: string;
    likendin?: string;
    github?: string;
    web_porfolio?: string;
    cv_uri?: string;
    status?: string;
    created_at?: number;
}